+++ assets/js/head/color-scheme.ts
@@ -1,3 +1,6 @@
+// @ts-expect-error — Hugo js.Build virtual module
+import params from '@params';
+
 /**
  * Color scheme initialization — runs synchronously in `<head>` before body rendering.
  *
@@ -5,10 +8,6 @@
  * - Read stored theme mode from localStorage, falling back to site default.
  * - Set `data-theme-mode` on `<html>` to prevent flash of wrong theme.
  */
-
-// @ts-expect-error — Hugo js.Build virtual module
-import params from '@params';
-
 (function () {
   const localStorage = window.localStorage
   const storedMode = localStorage?.getItem('theme-mode')
