+++ assets/js/lib/json-viewer.ts
@@ -5,10 +5,8 @@
  * - Apply theme-aware styling to json-viewer custom elements.
  * - Sync theme when it changes.
  */
-import { TypedEventBus } from '../core/event-bus'
+import { eventBus } from '../core/event-bus'
 import { isDarkMode } from '../utils'
-
-const eventBus = new TypedEventBus()
 
 function applyJsonViewerTheme(isDark: boolean) {
   document.querySelectorAll('json-viewer').forEach(($el: Element) => {
