+++ assets/js/modules/core.ts
@@ -1,3 +1,7 @@
+import type { CoreService } from '../core/tokens'
+import type { FixItConfig, MaskOverlayHandler } from '../types'
+import { getScrollTop, getThemeMode, isDarkMode } from '../utils'
+
 /**
  * Core module — shared state initialization and mask overlay management.
  *
@@ -6,10 +10,6 @@
  * - Track theme mode (light/dark) and provide `isDark` / `themeMode` accessors.
  * - Manage mask overlay visibility for search and menu drawers.
  */
-import type { CoreService } from '../core/tokens'
-import type { FixItConfig, MaskOverlayHandler } from '../types'
-import { getScrollTop, getThemeMode, isDarkMode } from '../utils'
-
 export class CoreModule implements CoreService {
   readonly config: FixItConfig
   themeMode: string
