--- assets/js/modules/svg.ts
@@ -1,31 +1,0 @@
-/**
- * SVG module — inline SVG icon loading from data-svg-src attributes.
- *
- * Responsibilities:
- * - Fetch and inline SVG icons referenced by `data-svg-src` attributes.
- * - Cache fetched SVGs to avoid duplicate requests.
- */
-import type { SvgService } from '../core/tokens'
-
-export class SvgModule implements SvgService {
-  /** Fetch and inline SVG icons referenced by `data-svg-src` attributes. */
-  initSVGIcon() {
-    document.querySelectorAll<HTMLElement>('[data-svg-src]').forEach(($icon) => {
-      fetch($icon.dataset.svgSrc!)
-        .then(response => response.text())
-        .then((svg) => {
-          const $temp = document.createElement('div')
-          $temp.insertAdjacentHTML('afterbegin', svg)
-          const $svg = $temp.firstChild as SVGElement
-          $svg.dataset.svgSrc = $icon.dataset.svgSrc
-          $svg.classList.add('icon')
-          const $titleElements = $svg.getElementsByTagName('title')
-          $titleElements.length && $svg.removeChild($titleElements[0])
-          $icon.parentElement!.replaceChild($svg, $icon)
-        })
-        .catch((err) => {
-          console.error(err)
-        })
-    })
-  }
-}
