+++ assets/js/modules/theme.ts
@@ -1,3 +1,6 @@
+import type { CoreService, ThemeService } from '../core/tokens'
+import { eventBus } from '../core/event-bus'
+
 /**
  * Theme module — color scheme switching and theme-color meta tag management.
  *
@@ -6,15 +9,32 @@
  * - Update `<meta name="theme-color">` based on current scheme.
  * - Persist user preference to localStorage.
  */
-
-import type { TypedEventBus } from '../core/event-bus'
-import type { CoreService, ThemeService } from '../core/tokens'
-
 export class ThemeModule implements ThemeService {
-  constructor(
-    private readonly core: CoreService,
-    private readonly bus: TypedEventBus,
-  ) {}
+  private readonly mql = window.matchMedia('(prefers-color-scheme: dark)')
+
+  constructor(private readonly core: CoreService) {}
+
+  /**
+   * Apply a theme mode and emit the `fixit:switch-theme` event.
+   * @param mode - `'auto'`, `'light'`, or `'dark'`.
+   * @param persist - Whether to save the choice to localStorage (default: `true`).
+   */
+  setThemeMode(mode: string, persist = true) {
+    const prevIsDark = this.core.isDark
+    this.core.themeMode = mode
+    document.documentElement.dataset.themeMode = mode
+    this.core.isDark = mode === 'auto' ? this.mql.matches : mode === 'dark'
+
+    if (persist) {
+      window.localStorage?.setItem('theme-mode', mode)
+    }
+
+    eventBus.emit('fixit:switch-theme', {
+      isDark: this.core.isDark,
+      mode,
+      isChanged: prevIsDark !== this.core.isDark,
+    })
+  }
 
   /** Sync the `<meta name="theme-color">` tag with the current color scheme. */
   initThemeColor() {
@@ -24,7 +44,7 @@ export class ThemeModule implements ThemeService {
     const applyThemeColor = (isDark: boolean) => {
       $meta.content = isDark ? $meta.dataset.dark! : $meta.dataset.light!
     }
-    this.bus.on('fixit:switch-theme', ({ detail }) => {
+    eventBus.on('fixit:switch-theme', ({ detail }) => {
       if (!detail.isChanged)
         return
       applyThemeColor(detail.isDark)
@@ -34,39 +54,22 @@ export class ThemeModule implements ThemeService {
 
   /** Initialize the theme switch button cycle and system preference listener. */
   initSwitchTheme() {
-    const mql = window.matchMedia('(prefers-color-scheme: dark)')
     const modes = ['auto', 'light', 'dark'] as const
-    const applyThemeMode = (mode: string, persist = true) => {
-      const prevIsDark = this.core.isDark
-      this.core.themeMode = mode
-      document.documentElement.dataset.themeMode = mode
-      this.core.isDark = mode === 'auto' ? mql.matches : mode === 'dark'
-
-      if (persist) {
-        window.localStorage?.setItem('theme-mode', mode)
-      }
-
-      this.bus.emit('fixit:switch-theme', {
-        isDark: this.core.isDark,
-        mode,
-        isChanged: prevIsDark !== this.core.isDark,
-      })
-    }
 
     document.querySelectorAll('.theme-switch').forEach(($themeSwitch: Element) => {
       $themeSwitch.addEventListener('click', () => {
         const currentIndex = modes.indexOf(this.core.themeMode as typeof modes[number])
         const nextMode = modes[(currentIndex + 1) % modes.length]
-        applyThemeMode(nextMode)
+        this.setThemeMode(nextMode)
       }, false)
     })
 
-    mql.addEventListener('change', (e: MediaQueryListEvent) => {
+    this.mql.addEventListener('change', (e: MediaQueryListEvent) => {
       if (this.core.themeMode !== 'auto')
         return
       const prevIsDark = this.core.isDark
       this.core.isDark = e.matches
-      this.bus.emit('fixit:switch-theme', {
+      eventBus.emit('fixit:switch-theme', {
         isDark: this.core.isDark,
         mode: 'auto',
         isChanged: prevIsDark !== this.core.isDark,
